requirement.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bilibili-api-python==16.1.1
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
lxml==4.9.3
multidict==6.0.4
Pillow==10.0.1
pyasn1==0.5.0
pycryptodomex==3.19.0
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pytz==2023.3.post1
pyuic5-tool==0.0.1
PyYAML==6.0.1
qrcode==7.4.2
qrcode-terminal==0.8
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
yarl==1.9.2
